pyproject.toml
[project]
name = "data-analyst-agent"
version = "0.1.0"
description = "Strands-based data analyst agent that authors AWS Glue jobs and Athena queries, served via FastAPI behind a Chainlit chat frontend."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # AWS SDKs + data wrangling
    "awswrangler>=3.0.0",
    "boto3>=1.42.54",
    "pandas<3.0.0",
    "scikit-learn>=1.8.0",

    # Strands agent + MCP tooling
    "strands-agents>=1.34.0",
    "strands-agents-tools>=0.2.21",
    "bedrock-agentcore>=1.3.1",  # AgentCoreCodeInterpreter sandbox client
    "mcp>=1.26.0",
    "awslabs-aws-dataprocessing-mcp-server>=0.1.29",
    "awslabs-aws-api-mcp-server>=1.3.33",

    # Observability
    "openinference-instrumentation-bedrock>=0.1.35",

    # FastAPI agent service
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",

    # Chainlit frontend + Postgres data layer
    "chainlit>=2.11.1",
    "plotly>=6.7.0",
    "asyncpg>=0.31.0",
    "sqlalchemy>=2.0.49",
    "greenlet>=3.5.0",

    # Misc
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "aws-cdk-lib>=2.251.0",
    "constructs>=10.6.0",
    "pytest>=8.4.2",
    "ruff>=0.15.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
